+++ __fixtures__/declared-gap/src/app/main.tsx
@@ -1,0 +1,5 @@
+import { App } from './App'
+import { crewRoutes } from '../modules/crews/routes'
+
+void App
+void crewRoutes